fix: refactor redirect utils
