+++ apps/arclight/src/lib/redirectUtils.ts
@@ -1,13 +1,15 @@
+import { Context } from 'hono'
+
 // Shared utilities for redirect routes
 
-export function setCorsHeaders(c: any) {
+export function setCorsHeaders(c: Context) {
   c.header('Access-Control-Allow-Origin', '*')
   c.header('Access-Control-Allow-Methods', 'GET, OPTIONS')
   c.header('Access-Control-Allow-Headers', '*')
   c.header('Access-Control-Expose-Headers', '*')
 }
 
-export function getClientIp(c: any): string | undefined {
+export function getClientIp(c: Context): string | undefined {
   const forwardedFor = c.req.header('x-forwarded-for')
   if (forwardedFor) {
     return forwardedFor.split(',')[0].trim()
